requirements.txt
Django==4.2.1
gTTS==2.3.2
openai==0.27.8
openai_whisper==20230314
python-dotenv==1.0.0
